requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
install==1.3.5
passlib==1.7.4
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-decouple==3.6
python-dotenv==0.19.2
python-multipart==0.0.5
python-stdnum==1.17
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
